+++ SqExpress/SyntaxTreeOperations/Internal/ExprValueTypeAnalyzer.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Data.SqlTypes;
+﻿using System.Data.SqlTypes;
 using SqExpress.Syntax;
 using SqExpress.Syntax.Expressions;
 using SqExpress.Syntax.Functions;
@@ -25,44 +24,48 @@ namespace SqExpress.SyntaxTreeOperations.Internal
 
     internal class ExprValueTypeAnalyzer<TRes, TCtx> : IExprValueVisitor<TRes, ExprValueTypeAnalyzerCtx<TRes, TCtx>>, IExprTypeVisitor<TRes, ExprValueTypeAnalyzerCtx<TRes, TCtx>>
     {
+        public static readonly ExprValueTypeAnalyzer<TRes, TCtx> Instance = new ExprValueTypeAnalyzer<TRes, TCtx>();
+
+        private ExprValueTypeAnalyzer() { }
+
         public TRes VisitExprInt32Literal(ExprInt32Literal exprInt32Literal, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
         {
-            return ctx.ValueVisitor.VisitInt32(ctx.Ctx);
+            return ctx.ValueVisitor.VisitInt32(ctx.Ctx, !exprInt32Literal.Value.HasValue);
         }
 
         public TRes VisitExprGuidLiteral(ExprGuidLiteral exprGuidLiteral, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
         {
-            return ctx.ValueVisitor.VisitGuid(ctx.Ctx);
+            return ctx.ValueVisitor.VisitGuid(ctx.Ctx, !exprGuidLiteral.Value.HasValue);
         }
 
         public TRes VisitExprStringLiteral(ExprStringLiteral stringLiteral, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
         {
-            return ctx.ValueVisitor.VisitString(ctx.Ctx, stringLiteral.Value?.Length, false);
+            return ctx.ValueVisitor.VisitString(ctx.Ctx, stringLiteral.Value == null, stringLiteral.Value?.Length, false);
         }
 
         public TRes VisitExprDateTimeLiteral(ExprDateTimeLiteral dateTimeLiteral, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
         {
-            return ctx.ValueVisitor.VisitDateTime(ctx.Ctx);
+            return ctx.ValueVisitor.VisitDateTime(ctx.Ctx, !dateTimeLiteral.Value.HasValue);
         }
 
         public TRes VisitExprBoolLiteral(ExprBoolLiteral boolLiteral, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
         {
-            return ctx.ValueVisitor.VisitBool(ctx.Ctx);
+            return ctx.ValueVisitor.VisitBool(ctx.Ctx, !boolLiteral.Value.HasValue);
         }
 
         public TRes VisitExprInt64Literal(ExprInt64Literal int64Literal, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
         {
-            return ctx.ValueVisitor.VisitInt64(ctx.Ctx);
+            return ctx.ValueVisitor.VisitInt64(ctx.Ctx, !int64Literal.Value.HasValue);
         }
 
         public TRes VisitExprByteLiteral(ExprByteLiteral byteLiteral, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
         {
-            return ctx.ValueVisitor.VisitByte(ctx.Ctx);
+            return ctx.ValueVisitor.VisitByte(ctx.Ctx, !byteLiteral.Value.HasValue);
         }
 
         public TRes VisitExprInt16Literal(ExprInt16Literal int16Literal, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
         {
-            return ctx.ValueVisitor.VisitInt16(ctx.Ctx);
+            return ctx.ValueVisitor.VisitInt16(ctx.Ctx, !int16Literal.Value.HasValue);
         }
 
         public TRes VisitExprDecimalLiteral(ExprDecimalLiteral decimalLiteral, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
@@ -75,27 +78,27 @@ namespace SqExpress.SyntaxTreeOperations.Internal
                 precisionScale = new DecimalPrecisionScale(sd.Precision, sd.Scale);
             }
 
-            return ctx.ValueVisitor.VisitDecimal(ctx.Ctx, precisionScale);
+            return ctx.ValueVisitor.VisitDecimal(ctx.Ctx, !decimalLiteral.Value.HasValue, precisionScale);
         }
 
         public TRes VisitExprDoubleLiteral(ExprDoubleLiteral doubleLiteral, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
         {
-            return ctx.ValueVisitor.VisitDouble(ctx.Ctx);
+            return ctx.ValueVisitor.VisitDouble(ctx.Ctx, !doubleLiteral.Value.HasValue);
         }
 
         public TRes VisitExprByteArrayLiteral(ExprByteArrayLiteral byteArrayLiteral, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
         {
-            return ctx.ValueVisitor.VisitByteArray(ctx.Ctx, byteArrayLiteral.Value?.Count, false);
+            return ctx.ValueVisitor.VisitByteArray(ctx.Ctx, byteArrayLiteral.Value == null, byteArrayLiteral.Value?.Count, false);
         }
 
         public TRes VisitExprNull(ExprNull exprNull, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
         {
-            return ctx.ValueVisitor.VisitAny(ctx.Ctx);
+            return ctx.ValueVisitor.VisitAny(ctx.Ctx, true);
         }
 
         public TRes VisitExprUnsafeValue(ExprUnsafeValue exprUnsafeValue, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
         {
-            return ctx.ValueVisitor.VisitAny(ctx.Ctx);
+            return ctx.ValueVisitor.VisitAny(ctx.Ctx, null);
         }
 
         public TRes VisitExprSum(ExprSum exprSum, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
@@ -125,12 +128,12 @@ namespace SqExpress.SyntaxTreeOperations.Internal
 
         public TRes VisitExprStringConcat(ExprStringConcat exprStringConcat, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
         {
-            return ctx.ValueVisitor.VisitString(ctx.Ctx, null, false);
+            return ctx.ValueVisitor.VisitString(ctx.Ctx, null ,null, false);
         }
 
         public TRes VisitExprScalarFunction(ExprScalarFunction exprScalarFunction, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
         {
-            return ctx.ValueVisitor.VisitAny(ctx.Ctx);
+            return ctx.ValueVisitor.VisitAny(ctx.Ctx, null);
         }
 
         public TRes VisitExprCase(ExprCase exprCase, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
@@ -155,17 +158,17 @@ namespace SqExpress.SyntaxTreeOperations.Internal
 
         public TRes VisitExprGetDate(ExprGetDate exprGetDate, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
         {
-            return ctx.ValueVisitor.VisitDateTime(ctx.Ctx);
+            return ctx.ValueVisitor.VisitDateTime(ctx.Ctx, false);
         }
 
         public TRes VisitExprGetUtcDate(ExprGetUtcDate exprGetUtcDate, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
         {
-            return ctx.ValueVisitor.VisitDateTime(ctx.Ctx);
+            return ctx.ValueVisitor.VisitDateTime(ctx.Ctx, false);
         }
 
         public TRes VisitExprDateAdd(ExprDateAdd exprDateAdd, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
         {
-            return ctx.ValueVisitor.VisitDateTime(ctx.Ctx);
+            return ctx.ValueVisitor.VisitDateTime(ctx.Ctx, null);
         }
 
         public TRes VisitExprColumn(ExprColumn exprColumn, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
@@ -174,7 +177,7 @@ namespace SqExpress.SyntaxTreeOperations.Internal
             {
                 return tc.SqlType.Accept(this, ctx);
             }
-            return ctx.ValueVisitor.VisitAny(ctx.Ctx);
+            return ctx.ValueVisitor.VisitAny(ctx.Ctx, null);
         }
 
         public TRes VisitExprCast(ExprCast exprCast, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
@@ -182,74 +185,76 @@ namespace SqExpress.SyntaxTreeOperations.Internal
             return exprCast.SqlType.Accept(this, ctx);
         }
 
-        public TRes VisitExprTypeBoolean(ExprTypeBoolean exprTypeBoolean, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
+        //Implementation to analyze in "VisitExprCast" and "VisitExprColumn"
+
+        TRes IExprTypeVisitor<TRes, ExprValueTypeAnalyzerCtx<TRes, TCtx>>.VisitExprTypeBoolean(ExprTypeBoolean exprTypeBoolean, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
         {
-            return ctx.ValueVisitor.VisitBool(ctx.Ctx);
+            return ctx.ValueVisitor.VisitBool(ctx.Ctx, null);
         }
 
-        public TRes VisitExprTypeByte(ExprTypeByte exprTypeByte, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
+        TRes IExprTypeVisitor<TRes, ExprValueTypeAnalyzerCtx<TRes, TCtx>>.VisitExprTypeByte(ExprTypeByte exprTypeByte, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
         {
-            return ctx.ValueVisitor.VisitByte(ctx.Ctx);
+            return ctx.ValueVisitor.VisitByte(ctx.Ctx, null);
         }
 
-        public TRes VisitExprTypeByteArray(ExprTypeByteArray exprTypeByte, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
+        TRes IExprTypeVisitor<TRes, ExprValueTypeAnalyzerCtx<TRes, TCtx>>.VisitExprTypeByteArray(ExprTypeByteArray exprTypeByte, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
         {
-            return ctx.ValueVisitor.VisitByteArray(ctx.Ctx, exprTypeByte.Size, false);
+            return ctx.ValueVisitor.VisitByteArray(ctx.Ctx, null, exprTypeByte.Size, false);
         }
 
-        public TRes VisitExprTypeFixSizeByteArray(ExprTypeFixSizeByteArray exprTypeFixSizeByteArray, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
+        TRes IExprTypeVisitor<TRes, ExprValueTypeAnalyzerCtx<TRes, TCtx>>.VisitExprTypeFixSizeByteArray(ExprTypeFixSizeByteArray exprTypeFixSizeByteArray, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
         {
-            return ctx.ValueVisitor.VisitByteArray(ctx.Ctx, exprTypeFixSizeByteArray.Size, true);
+            return ctx.ValueVisitor.VisitByteArray(ctx.Ctx, null, exprTypeFixSizeByteArray.Size, true);
         }
 
-        public TRes VisitExprTypeInt16(ExprTypeInt16 exprTypeInt16, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
+        TRes IExprTypeVisitor<TRes, ExprValueTypeAnalyzerCtx<TRes, TCtx>>.VisitExprTypeInt16(ExprTypeInt16 exprTypeInt16, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
         {
-            return ctx.ValueVisitor.VisitInt16(ctx.Ctx);
+            return ctx.ValueVisitor.VisitInt16(ctx.Ctx, null);
         }
 
-        public TRes VisitExprTypeInt32(ExprTypeInt32 exprTypeInt32, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
+        TRes IExprTypeVisitor<TRes, ExprValueTypeAnalyzerCtx<TRes, TCtx>>.VisitExprTypeInt32(ExprTypeInt32 exprTypeInt32, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
         {
-            return ctx.ValueVisitor.VisitInt32(ctx.Ctx);
+            return ctx.ValueVisitor.VisitInt32(ctx.Ctx, null);
         }
 
-        public TRes VisitExprTypeInt64(ExprTypeInt64 exprTypeInt64, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
+        TRes IExprTypeVisitor<TRes, ExprValueTypeAnalyzerCtx<TRes, TCtx>>.VisitExprTypeInt64(ExprTypeInt64 exprTypeInt64, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
         {
-            return ctx.ValueVisitor.VisitInt64(ctx.Ctx);
+            return ctx.ValueVisitor.VisitInt64(ctx.Ctx, null);
         }
 
-        public TRes VisitExprTypeDecimal(ExprTypeDecimal exprTypeDecimal, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
+        TRes IExprTypeVisitor<TRes, ExprValueTypeAnalyzerCtx<TRes, TCtx>>.VisitExprTypeDecimal(ExprTypeDecimal exprTypeDecimal, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
         {
-            return ctx.ValueVisitor.VisitDecimal(ctx.Ctx, exprTypeDecimal.PrecisionScale);
+            return ctx.ValueVisitor.VisitDecimal(ctx.Ctx, null, exprTypeDecimal.PrecisionScale);
         }
 
-        public TRes VisitExprTypeDouble(ExprTypeDouble exprTypeDouble, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
+        TRes IExprTypeVisitor<TRes, ExprValueTypeAnalyzerCtx<TRes, TCtx>>.VisitExprTypeDouble(ExprTypeDouble exprTypeDouble, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
         {
-            return ctx.ValueVisitor.VisitDouble(ctx.Ctx);
+            return ctx.ValueVisitor.VisitDouble(ctx.Ctx, null);
         }
 
-        public TRes VisitExprTypeDateTime(ExprTypeDateTime exprTypeDateTime, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
+        TRes IExprTypeVisitor<TRes, ExprValueTypeAnalyzerCtx<TRes, TCtx>>.VisitExprTypeDateTime(ExprTypeDateTime exprTypeDateTime, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
         {
-            return ctx.ValueVisitor.VisitDateTime(ctx.Ctx);
+            return ctx.ValueVisitor.VisitDateTime(ctx.Ctx, null);
         }
 
-        public TRes VisitExprTypeGuid(ExprTypeGuid exprTypeGuid, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
+        TRes IExprTypeVisitor<TRes, ExprValueTypeAnalyzerCtx<TRes, TCtx>>.VisitExprTypeGuid(ExprTypeGuid exprTypeGuid, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
         {
-            return ctx.ValueVisitor.VisitGuid(ctx.Ctx);
+            return ctx.ValueVisitor.VisitGuid(ctx.Ctx, null);
         }
 
-        public TRes VisitExprTypeString(ExprTypeString exprTypeString, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
+        TRes IExprTypeVisitor<TRes, ExprValueTypeAnalyzerCtx<TRes, TCtx>>.VisitExprTypeString(ExprTypeString exprTypeString, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
         {
-            return ctx.ValueVisitor.VisitString(ctx.Ctx, exprTypeString.Size, false);
+            return ctx.ValueVisitor.VisitString(ctx.Ctx, null, exprTypeString.Size, false);
         }
 
-        public TRes VisitExprTypeFixSizeString(ExprTypeFixSizeString exprTypeFixSizeString, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
+        TRes IExprTypeVisitor<TRes, ExprValueTypeAnalyzerCtx<TRes, TCtx>>.VisitExprTypeFixSizeString(ExprTypeFixSizeString exprTypeFixSizeString, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
         {
-            return ctx.ValueVisitor.VisitString(ctx.Ctx, exprTypeFixSizeString.Size, true);
+            return ctx.ValueVisitor.VisitString(ctx.Ctx, null, exprTypeFixSizeString.Size, true);
         }
 
-        public TRes VisitExprTypeXml(ExprTypeXml exprTypeXml, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
+        TRes IExprTypeVisitor<TRes, ExprValueTypeAnalyzerCtx<TRes, TCtx>>.VisitExprTypeXml(ExprTypeXml exprTypeXml, ExprValueTypeAnalyzerCtx<TRes, TCtx> ctx)
         {
-            return ctx.ValueVisitor.VisitXml(ctx.Ctx);
+            return ctx.ValueVisitor.VisitXml(ctx.Ctx, null);
         }
     }
 }